+++ index.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="styles.css">
     <link rel="preconnect" href="https://fonts.gstatic.com">
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap" rel="stylesheet">
-    <script src="scripts/main-scripts.js"></script>
+    <script type="module" src="scripts/main-scripts.js"></script>
 </head>
 
 <body>
