+++ scripts/services/animation-service.js
@@ -1,0 +1,11 @@
+export class AnimationService {
+    #requestAnimationFrameId;
+
+    getAnimationId() {
+        return this.#requestAnimationFrameId;
+    }
+
+    setAnimationId(id) {
+        this.#requestAnimationFrameId = id;
+    }
+}
