+++ scripts/services/media-service.js
@@ -1,0 +1,13 @@
+export class MediaService {
+    #listeners = [];
+
+    registerMediaPlaying(listener) {
+        this.#listeners.push(listener);
+    }
+
+    notifyMediaPlaying(eventComp) {
+        for (let listener of this.#listeners) {
+            listener(eventComp);
+        }
+    }
+}
